% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S5-match-age-distribution.R
\name{age.match}
\alias{age.match}
\title{\enumerate{
\item Match age distribution of survey participants
}}
\usage{
age.match(mpx1b, data1980control, method)
}
\arguments{
\item{mpx1b}{MPX 2007 survey dataset}

\item{data1980control}{Control responses from the 1980 MPX survey}

\item{method}{Controls how ages are matched or assembled from the 2000s survey. Default is 'matched' (also takes 'all<27' or 'unif<27')}
}
\value{
A vector of rows to sample to match the age distribution of the 1980s survey
}
\description{
The 1982 survey data come from the Control group of unaffected individuals, roughly age-matched to individuals who were seropositive for exposure to MPXV.  Because of the distribution of susceptibility at this time, the composition of the 1982 respondents are strongly skewed to children under 10 years of age.  In contrast, the 2007 survey includes individuals of all ages, who might exhibit quite different patterns of animal contact from children.  To control for the age of respondents, and hence to arrive at the most accurate estimate for the influence of changes in the wildlife community on the animal-human contact rates in this system, we subsampled the 2007 survey respondents to match the distribution of ages in the 1982 respondents.
}
\examples{
age.match(mpx1b, data1980control, method = 'matched')
}
\keyword{responses}
