% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2-aggregate-contact-types.R
\name{contact.type.max}
\alias{contact.type.max}
\title{\enumerate{
\item Aggregate contact types by species
}}
\usage{
contact.type.max(mpx1b, orig.animal.names)
}
\arguments{
\item{mpx1b}{MPX 2007 survey dataset}

\item{orig.animal.names}{The original animal names given in the 2007 survey}
}
\value{
A data frame with the maximum contact value for each species for each surveyed individual
}
\description{
The 2007 survey included separate questions on 10 different types of contacts (hunt, found dead, butcher, skin, cook, eat, eat raw, play with, scratched by, bitten by).  Because we aimed to make quantitative comparisons with the 1982 survey, which did not collect separate frequencies for different types of contact, we needed to aggregate the contact types to a single representative frequency.  As a conservative assumption, we assumed that different types of contact with a given animal species were likely to occur on the same day, so contact frequencies should not be summed.  Instead, we took the maximum frequency reported across all contact types within a species, and used this as the contact frequency for that species.
}
\examples{
contact.type.max(mpx1b, orig.animal.names)
}
