% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1980s.R
\name{data1980.calc.contact}
\alias{data1980.calc.contact}
\title{Analyze 1980 dataset}
\usage{
data1980.calc.contact(
  i,
  data1980control,
  case = "resample",
  squirrel = FALSE,
  antelope = FALSE,
  rodents = FALSE
)
}
\arguments{
\item{i}{Set to 1 if running function independently. Used for replication of the function in apply statement.}

\item{data1980control}{Control responses from the 1980 MPX survey}

\item{case}{Defines how frequencies not included in the top three are analyzed. Default is 'resample', which dictates that non-top 3 frequencies are uniformly sampled from 0 to the #3 frequency. Also takes 'lower', which defines all non-top 3 frequencies as 0.}

\item{squirrel}{Indicates whether or not to return the raw squirrel frequency data. Default is FALSE.}

\item{antelope}{Indicates whether or not to return the raw antelope frequency data. Default is FALSE.}

\item{rodents}{Indicates whether or not to return the raw rodent frequency data. Default is FALSE.}
}
\value{
A list containing the mean and standard deviation values by species group (rodent, NHP, antelope, pig, squirrel) for the 1980s dataset
}
\description{
Analyze 1980 dataset
}
\keyword{responses}
