% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/S4-standardize-how-less-common-groups-are-included.R
\name{replicate.1980}
\alias{replicate.1980}
\title{\enumerate{
\item Standardize how less commonly contacted groups are included
}}
\usage{
replicate.1980(total.contacts.by.group, case = "resample")
}
\arguments{
\item{total.contacts.by.group}{A group-level summary of contacts (generated via aggregate.species.groups())}

\item{case}{Determines how to summarize the 2000s data. Suitable values are: resample or lower.}
}
\value{
A data frame of the top 3 contacts by each surveyed individual
}
\description{
While the 2007 survey collected contact-frequency information for all species, Because the 1982 survey only collected data on the 3 most contacted groups. When full data were available, frequent contacts were often reported with more than the top three groups, and ignoring these missing values from the 1982 survey would result in biased contact estimates. To harmonize the 1982 and 2007 surveys, we assigned sampled values to animal groups that were not reported as the top three most-frequent contacts. Values were sampled from a uniform distribution from zero to the third-place contact rate.
}
\details{
To bring data from the two surveys onto equivalent footing, we applied the following consistent set of actions: 1) For 1982 responses, use data on all 3 groups reported.  For all other groups, not reported explicitly, the contact rate was unifromtly sampled from zero to the third place contact rate for that respondent. 2) For 2007 responses, use data from the 3 groups with highest reported contact frequencies.  In the event of a tie for the 3rd spot, choose randomly which to include. For all other groups, below the top 3, we used the same procedure as for 1982 survey data, to maximize comparability.
}
