% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.MPXfxns.2000s.R
\name{run.MPXfxns.2000s}
\alias{run.MPXfxns.2000s}
\title{Run all steps of 2000s analysis}
\usage{
run.MPXfxns.2000s(
  ii,
  mpx1b,
  data1980control,
  orig.animal.names,
  animal.group.names,
  animal.groups,
  case = "resample",
  method = "matched"
)
}
\arguments{
\item{ii}{Set to 1 if running function independently. Used for function iteration in apply statement.}

\item{mpx1b}{MPX 2007 survey dataset}

\item{data1980control}{Control responses from the 1980 MPX survey}

\item{orig.animal.names}{The original animal names given in the 2007 survey}

\item{animal.group.names}{Unique animal groups to match the 1980 MPX survey}

\item{animal.groups, }{Non-unique animal groups to which the orig.animal.names are assigned}

\item{case}{Default = 'lower'. Also takes 'upper' or 'midpoint'}

\item{method}{Default = 'matched'. Also takes 'all<27' or 'unif<27'}
}
\value{
A vector of rows to sample to match the age distribution of the 1980s survey
}
\description{
Run all steps of 2000s analysis
}
\keyword{responses}
