% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S1-convert-survey-responses-to-contact-freqs.R
\name{sample.contact.freq}
\alias{sample.contact.freq}
\title{\enumerate{
\item Convert 2007 survey responses to contact frequencies
}}
\usage{
sample.contact.freq(mpx1b)
}
\arguments{
\item{mpx1b}{MPX 2007 survey dataset}
}
\value{
A data frame with recoded survey values, which are sampled from triangular distributions
}
\description{
The 2007 survey had five categorical options for frequency of contacts:
}
\details{
\tabular{ll}{
   Code \tab Frequency \cr
   0 \tab Never \cr
   1 \tab Every day \cr
   2 \tab 1 time/month \cr
   3 \tab 2-3 times/month \cr
   4 \tab >4 times/month \cr
}


To convert these into quantitative data that could be compared with data
from 1982, we sampled contact frequencies from the following triangular
distributions:\tabular{lll}{
   Code \tab Frequency \tab Sampling distribution \cr
   0 \tab Never \tab 0 \cr
   1 \tab Every day \tab rtri(1, min = 20, max = 40, mode = 30) \cr
   2 \tab 1 time/month \tab rtri(1, min = 0.5, max = 1.5, mode = 1) \cr
   3 \tab 2-3 times/month \tab rtri(1, min = 1.5, max = 3.5, mode = 2.5) \cr
   4 \tab >4 times/month \tab rtri(1, min = 4, max = 20, mode = 12) \cr
}
}
\examples{
sample.contact.freq(mpx1b)
}
\keyword{responses}
