% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-aggregate-species-to-groups.R
\name{species.groups.sum}
\alias{species.groups.sum}
\title{\enumerate{
\item Aggregate species into groups, matching 1982 groups
}}
\usage{
species.groups.sum(mpx1b, animal.group.names, max.contacts.by.species)
}
\arguments{
\item{mpx1b}{MPX 2007 survey dataset}

\item{animal.group.names}{Animal groups to match the 1980 MPX survey}

\item{max.contacts.by.species}{The data frame output from contact.type.max}
}
\value{
A data frame with the contact values summed across species group
}
\description{
The 2007 survey included separate questions for 28 species of animal, including 12 species of non-human primates and 7 species of rodents.  To allow robust comparison with the coarser classifications used in the 1982 responses, we aggregated the species into four broad groups: antelope, rodent, NHP, and pig.  Because squirrels are the suspected reservoir of MPXV in the DRC, we conducted a separate analysis that focused on reported contacts with squirrels in particular.
}
\details{
Contact rates were summed for all species within a group.  Because 12 species of NHP were included in the 2007 questionnaire, whereas most responses to the 1982 survey simply said ‘singe’ or ‘XXX’, this led to inflation of the 2007 values.  There is not sufficient information in the two surveys to reconcile this difference, so we did not include NHP in our final analyses.
}
